requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.1.0
email_validator==2.2.0
itsdangerous==2.2.0
Werkzeug==3.1.3
Jinja2==3.1.6
SQLAlchemy==2.0.40
requests==2.32.3
openai==1.75.0
sentence-transformers==2.2.2
scikit-learn==1.6.1
joblib==1.4.2
numpy==1.26.4
gunicorn==21.2.0